实现用户信息查询功能

+++ OnlineTeachPlat/static/libs/js/oltp.js
@@ -9,6 +9,42 @@ $(document).ready(function(){
   	});
 
 /************************************
+*     查找用户信息
+*     admin_index.html --> views.admin
+**************************************/
+$("#modal_button_search").click('shown',function(e){
+    // 判断是否有输入信息
+    search_user = $("input[id=modal_input_search]").val();
+    if(search_user == ""){
+      alert("请输入用户名！！！")
+    }
+    else{
+      data = {
+        type:"search",
+        search_name:search_user
+      }
+      $.post("/oltp/admin/",data,function(data1,status){
+          json = JSON.parse(data1)
+          if(json.exist == "1"){
+              $('#myModal').modal('hide')
+              tr = document.getElementsByClassName("list")
+              for(var i=1;i<tr.length;i++){
+                    tr[i].style.display="none";
+              }
+              tr[0].children.td_username.innerText = json.username
+              tr[0].children.td_email.innerText = json.email
+              tr[0].children.td_group.innerText = json.group
+              tr[0].children.td_active.innerText = json.status
+        }
+        else {
+            alert("用户不存在");
+        }
+      })
+    }
+
+});
+
+/************************************
 *     更改用户
 *     admin_index.html --> views.admin
 **************************************/
@@ -44,12 +80,34 @@ $(function (){ // 注意dialog需要外层function包裹
               }
             
              $.post("/oltp/admin/",data,function(data1,status){
-                  // $(this).dialog("close");
                   alert("Data: " + data1 + "\nStatus: " + status);
-                  $("td").removeClass("selected");
             })
-             $(this).dialog("close");
-            
+            $(this).dialog("close");
+          
+            // 修改页面显示
+            selecte = document.getElementsByClassName("selected")
+            if (update_email == ""){
+              selecte.td_email.innerText = document.getElementById("dialog_update_email").placeholder
+            }
+            else{
+              selecte.td_email.innerText = update_email
+            }
+            if (update_status == ""){
+              selecte.td_active.innerText = document.getElementById("dialog_update_status").placeholder
+            }
+            else{
+              selecte.td_active.innerText = update_status
+            }
+
+            if (update_group == ""){
+              selecte.td_group.innerText = document.getElementById("dialog_update_group").placeholder
+            }
+            else{
+              selecte.td_group.innerText = update_group
+            }
+
+            $("td").removeClass("selected");  
+
       }
     }
 
@@ -77,7 +135,7 @@ $(function (){ // 注意dialog需要外层function包裹
    update = document.getElementById("dialog_update")
    select = $('.selected')
    var num=$('.selected').length
-   if (num == 4){
+   if (num == 4){ // 控制用户选择行数
       update.innerHTML ='\
         <h3 id="dialog_update_username" align="left">'+select[0].innerText+'</h3> \
         <div class="input-group"> \
@@ -182,32 +240,39 @@ $(function (){ // 注意dialog需要外层function包裹
 
   $("#id_delete").click(function(e){
   	// console.log(e);
-  	$('td').nextAll('.selected').parent().hide()
 
-  	var num=$('.selected').length
-  	var j=0;
+    select = $('.selected')
+    var num=$('.selected').length
+    if (num>3){
+          	$('td').nextAll('.selected').parent().hide()
 
-  	var data = {type:"delete"};
-  	var str  = ""
-  	var sp = ","
-  	// 取数据,依赖于每行有4列来取每行首列
-  	for (var i=0;i<num;i=i+4)
-  	{
-  		a = $(".selected:eq("+i+")").text()
-  		j = j + 1;
-  		num = num - 1;
-  		str = str.concat(a).concat(sp)
-  		// data[j] = $(".selected:eq("+i+")").text()
-  	}
-  	data["usernames"] = str
+          	var num=$('.selected').length
+          	var j=0;
 
-  	// console.log(str)
-  	// console.log(data)
-	// 后台删
-	$.post("/oltp/admin/",data,function(data1,status){
-		alert("Data: " + data1 + "\nStatus: " + status);
-	})
+          	var data = {type:"delete"};
+          	var str  = ""
+          	var sp = ","
+          	// 取数据,依赖于每行有4列来取每行首列
+          	for (var i=0;i<num;i=i+4)
+          	{
+          		a = $(".selected:eq("+i+")").text()
+          		j = j + 1;
+          		num = num - 1;
+          		str = str.concat(a).concat(sp)
+          		// data[j] = $(".selected:eq("+i+")").text()
+          	}
+          	data["usernames"] = str
 
-  });
+          // console.log(str)
+          // console.log(data)
+        	// 后台删
+        	$.post("/oltp/admin/",data,function(data1,status){
+        		alert("Data: " + data1 + "\nStatus: " + status);
+        	})
+        }
+        else{
+            alert("请选择至少选择一行数据") 
+        }
+    });
 
 });
